requirements.txt
aiomysql
python-dotenv
py-cord